requirements.text.txt
streamlit
pandas
numpy
opencv-python
Pillow
scipy